requirements.txt
Flask==2.3.2
matplotlib==3.8.1
pandas==2.0.3
pg8000==1.29.8
python-dotenv==1.0.0
bcrypt==4.0.1
gunicorn
SQLAlchemy==2.0.23
sendgrid==6.11.0